backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Chinmayee Panda <andolasoft.user206@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
fastapi-users = {extras = ["oauth", "sqlalchemy"], version = "^13.0.0"}
pydantic-settings = "^2.2.1"
aiomysql = "^0.2.0"
sqlacodegen-v2 = "^0.1.4"
mysqlclient = "^2.2.4"
fastapi-query = "^0.1a0"
loguru = "^0.7.2"
validators = "^0.28.0"
jinja2 = "^3.1.3"
mailtrap = "^2.0.1"
aioboto3 = "^12.4.0"
pycountry = "^23.12.11"
pytz = "^2024.1"
currencies = "^2020.12.12"
httpx = "^0.27.0"
forex-python = "^1.8"
redis-om = "^0.3.0"
redis = "^5.0.4"
fastapi-redis-rate-limiter = "^1.0.1"
aiofiles = "^23.2.1"
json-delta = "^2.0.2"
anytree = "^2.12.1"
treelib = "^1.7.0"
pandas = "^2.2.2"
beautifulsoup4 = "^4.12.3"
xhtml2pdf = "^0.2.16"
firebase-admin = "^6.5.0"
twilio = "^9.3.7"
python-jose = "^3.3.0"
pyotp = "^2.9.0"
mysql-connector = "^2.2.9"
validate-email = "^1.3"
python-barcode = "^0.15.1"
pillow = "^11.1.0"
whoosh = "^2.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
